+++ src/loaders.rs
@@ -1,6 +1,0 @@
-// Copyright 2020 Peter Williams <peter@newton.cx> and collaborators
-// Licensed under the MIT License.
-
-//! Auto-loading project information from the repository.
-
-pub mod cargo;