+++ src/rewriters.rs
@@ -5,8 +5,6 @@
 
 use crate::{app::AppSession, errors::Result, repository::ChangeList};
 
-pub mod cargo;
-
 /// A trait for something that can perform some kind of metadata rewriting.
 pub trait Rewriter: std::fmt::Debug {
     fn rewrite(&self, app: &AppSession, changes: &mut ChangeList) -> Result<()>;